requirements.txt
streamlit==1.28.2
pillow==10.1.0
pandas==2.1.1
joblib==1.3.2